requirements.txt
streamlit
torch
matplotlib
numpy
pillow
opencv-python